+++ priv/test_repo/migrations/20260127000000_add_durable_objects.exs
@@ -1,11 +1,6 @@
 defmodule DurableObject.TestRepo.Migrations.AddDurableObjects do
   use Ecto.Migration
 
-  def up do
-    DurableObject.Migration.up(version: 1)
-  end
-
-  def down do
-    DurableObject.Migration.down(version: 1)
-  end
+  def up, do: DurableObject.Migration.up()
+  def down, do: DurableObject.Migration.down()
 end